+++ Cargo.toml
@@ -30,9 +30,11 @@ members = [
     "dns-server",
     "dns-server-api",
     "end-to-end-tests",
+    "gateway",
+    "gateway-api",
     "gateway-cli",
     "gateway-test-utils",
-    "gateway",
+    "gateway-types",
     "illumos-utils",
     "installinator-api",
     "installinator-common",
@@ -132,9 +134,11 @@ default-members = [
     "dns-server-api",
     # Do not include end-to-end-tests in the list of default members, as its
     # tests only work on a deployed control plane.
+    "gateway",
+    "gateway-api",
     "gateway-cli",
     "gateway-test-utils",
-    "gateway",
+    "gateway-types",
     "illumos-utils",
     "installinator-api",
     "installinator-common",
@@ -313,10 +317,12 @@ flume = "0.11.0"
 foreign-types = "0.3.2"
 fs-err = "2.11.0"
 futures = "0.3.30"
+gateway-api = { path = "gateway-api" }
 gateway-client = { path = "clients/gateway-client" }
 gateway-messages = { git = "https://github.com/oxidecomputer/management-gateway-service", rev = "c85a4ca043aaa389df12aac5348d8a3feda28762", default-features = false, features = ["std"] }
 gateway-sp-comms = { git = "https://github.com/oxidecomputer/management-gateway-service", rev = "c85a4ca043aaa389df12aac5348d8a3feda28762" }
 gateway-test-utils = { path = "gateway-test-utils" }
+gateway-types = { path = "gateway-types" }
 gethostname = "0.4.3"
 glob = "0.3.1"
 guppy = "0.17.5"
