+++ dev-tools/openapi-manager/Cargo.toml
@@ -17,6 +17,7 @@ clap.workspace = true
 dns-server-api.workspace = true
 dropshot.workspace = true
 fs-err.workspace = true
+gateway-api.workspace = true
 indent_write.workspace = true
 installinator-api.workspace = true
 nexus-internal-api.workspace = true
