+++ dev-tools/openapi-manager/src/spec.rs
@@ -15,17 +15,6 @@ use openapiv3::OpenAPI;
 pub fn all_apis() -> Vec<ApiSpec> {
     vec![
         ApiSpec {
-            title: "CockroachDB Cluster Admin API",
-            version: "0.0.1",
-            description: "API for interacting with the Oxide control plane's \
-                CockroachDB cluster",
-            boundary: ApiBoundary::Internal,
-            api_description:
-                cockroach_admin_api::cockroach_admin_api_mod::stub_api_description,
-            filename: "cockroach-admin.json",
-            extra_validation: None,
-        },
-        ApiSpec {
             title: "Bootstrap Agent API",
             version: "0.0.1",
             description: "Per-sled API for setup and teardown",
@@ -33,6 +22,28 @@ pub fn all_apis() -> Vec<ApiSpec> {
             api_description:
                 bootstrap_agent_api::bootstrap_agent_api_mod::stub_api_description,
             filename: "bootstrap-agent.json",
+            extra_validation: None,
+        },
+        ApiSpec {
+            title: "CockroachDB Cluster Admin API",
+            version: "0.0.1",
+            description: "API for interacting with the Oxide \
+                control plane's CockroachDB cluster",
+            boundary: ApiBoundary::Internal,
+            api_description:
+                cockroach_admin_api::cockroach_admin_api_mod::stub_api_description,
+            filename: "cockroach-admin.json",
+            extra_validation: None,
+        },
+        ApiSpec {
+            title: "Oxide Management Gateway Service API",
+            version: "0.0.1",
+            description: "API for interacting with the Oxide \
+                control plane's gateway service",
+            boundary: ApiBoundary::Internal,
+            api_description:
+                gateway_api::gateway_api_mod::stub_api_description,
+            filename: "gateway.json",
             extra_validation: None,
         },
         ApiSpec {
