+++ gateway-types/src/component_details.rs
@@ -2,8 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-// Copyright 2023 Oxide Computer Company
-
 use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
