+++ gateway-types/src/lib.rs
@@ -2,8 +2,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-// Copyright 2022 Oxide Computer Company
+//! Common types for MGS.
 
-mod component_list;
-mod location_discovery;
-mod serial_console;
+pub mod caboose;
+pub mod component;
+pub mod component_details;
+pub mod host;
+pub mod ignition;
+pub mod rot;
+pub mod sensor;
+pub mod update;
