+++ gateway/Cargo.toml
@@ -14,8 +14,10 @@ camino.workspace = true
 clap.workspace = true
 dropshot.workspace = true
 futures.workspace = true
+gateway-api.workspace = true
 gateway-messages.workspace = true
 gateway-sp-comms.workspace = true
+gateway-types.workspace = true
 hex.workspace = true
 http.workspace = true
 hyper.workspace = true
