+++ sp-sim/src/gimlet.rs
@@ -36,6 +36,7 @@ use gateway_messages::SpStateV2;
 use gateway_messages::{version, MessageKind};
 use gateway_messages::{ComponentDetails, Message, MgsError, StartupOptions};
 use gateway_messages::{DiscoverResponse, IgnitionState, PowerState};
+use gateway_types::component::SpState;
 use slog::{debug, error, info, warn, Logger};
 use std::cell::Cell;
 use std::collections::HashMap;
@@ -104,8 +105,8 @@ impl Drop for Gimlet {
 
 #[async_trait]
 impl SimulatedSp for Gimlet {
-    async fn state(&self) -> omicron_gateway::http_entrypoints::SpState {
-        omicron_gateway::http_entrypoints::SpState::from(
+    async fn state(&self) -> SpState {
+        SpState::from(
             self.handler.as_ref().unwrap().lock().await.sp_state_impl(),
         )
     }
